+++ js-miniapp-sdk/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## CHANGELOG
 
+### In Progress (2021-XX-XX)
+- **Feature:** Support send message to contact. [See here](README.md#Send-message).
+
 ### 1.6.1 (2021-01-06)
 - **Fix:** Location permission support for iOS/Android Mini App SDK 2.6 and below.
 - **Fix:** Reject promise from `MiniApp.requestLocationPermission` when the user denies location custom permission.